requirements.txt
astor
requests
orjson
typing-extensions
tree-sitter
tree-sitter-python
pyyaml
coverage
pytest
responses
pynacl